\name{IBC}
\alias{IBC}
\title{
Information Based Criterion
}
\description{
Calculate the entropy of each variable of x
}
\usage{
IBC(x)
}
\arguments{
  \item{x}{ matrix including the SNP data
}
}

\details{
The SNPs data is categorical, and the entropy can be appropriate to quantify the information in SNP's variable, entropy is a measure of the uncertainty associated with a random variable.
}

\references{
Amiri, S., Ivo, D. (2017). Information Theoretic Approach for Genome-Wide Association Study of Parkinson's Disease.
}

\seealso{
See \code{\link{HFB}} for the application.
}
\keyword{Function}
