\name{IBC}
\alias{IBC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Information Based Criterion
}
\description{
Calculate the entropy of each variabled of x
}
\usage{
IBC(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ matrix including the SNP data
}
}

\details{
The SNPs data is categorical, and the entropy can be appropriate to quantify the information in SNP's variable, entropy is a measure of the uncertainty associated with a random variable.
}

\references{
Amiri, S., Ivo, D. (2015). Genome Wide-Association Study with application of Parkinson's Disease.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~
\seealso{
See \code{\link{HFB}} for the application.
}
\keyword{Function}
